Silence some compiler warnings
These were introduced in the tr/// changes in the series
merged in 240494d6992696a7a350217c131e1d5dc1444a0c

+++ op.c
@@ -7754,8 +7754,8 @@ S_pmtrans(pTHX_ OP *o, OP *expr, OP *repl)
 
         tbl->map[tbl->size] = del
                               ? (short) TR_DELETE
-                              : rlen
-                                ? final_map
+                              : (short) rlen
+                                ? (short) final_map
                                 : (short) TR_R_EMPTY;
 
         SvREFCNT_dec(t_invlist);
